`Removed unused error pages and components`
